9077: Move processPractitionerMapping records into its own file
