+++ shared/src/business/useCases/elasticsearch/processPractitionerMappingEntries.js
@@ -1,0 +1,83 @@
+const AWS = require('aws-sdk');
+const { flattenDeep } = require('lodash');
+
+export const processPractitionerMappingEntries = async ({
+  applicationContext,
+  practitionerMappingRecords,
+}) => {
+  if (!practitionerMappingRecords.length) return;
+
+  const indexCaseEntryForPractitionerMapping =
+    async practitionerMappingRecord => {
+      const practitionerMappingData =
+        practitionerMappingRecord.dynamodb.NewImage ||
+        practitionerMappingRecord.dynamodb.OldImage;
+      const caseRecords = [];
+
+      const caseMetadataWithCounsel = await applicationContext
+        .getPersistenceGateway()
+        .getCaseMetadataWithCounsel({
+          applicationContext,
+          docketNumber: practitionerMappingData.pk.S.substring('case|'.length),
+        });
+
+      const marshalledCase = AWS.DynamoDB.Converter.marshall(
+        caseMetadataWithCounsel,
+      );
+
+      caseRecords.push({
+        dynamodb: {
+          Keys: {
+            pk: {
+              S: practitionerMappingData.pk.S,
+            },
+            sk: {
+              S: practitionerMappingData.pk.S,
+            },
+          },
+          NewImage: {
+            ...marshalledCase,
+            case_relations: { name: 'case' },
+            entityName: { S: 'CaseDocketEntryMapping' },
+          }, // Create a mapping record on the docket-entry index for parent-child relationships
+        },
+        eventName: 'MODIFY',
+      });
+
+      caseRecords.push({
+        dynamodb: {
+          Keys: {
+            pk: {
+              S: practitionerMappingData.pk.S,
+            },
+            sk: {
+              S: practitionerMappingData.sk.S,
+            },
+          },
+          NewImage: marshalledCase,
+        },
+        eventName: 'MODIFY',
+      });
+
+      return caseRecords;
+    };
+
+  const indexRecords = await Promise.all(
+    practitionerMappingRecords.map(indexCaseEntryForPractitionerMapping),
+  );
+
+  const { failedRecords } = await applicationContext
+    .getPersistenceGateway()
+    .bulkIndexRecords({
+      applicationContext,
+      records: flattenDeep(indexRecords),
+    });
+
+  if (failedRecords.length > 0) {
+    applicationContext.logger.error(
+      'the practitioner mapping record that failed to index',
+      { failedRecords },
+    );
+    throw new Error('failed to index practitioner mapping records');
+  }
+};
