+++ shared/src/business/useCases/elasticsearch/processStreamUtilities.test.js
@@ -4,7 +4,6 @@ const {
   isPractitionerMappingRemoveRecord,
   partitionRecords,
   processOtherEntries,
-  processPractitionerMappingEntries,
 } = require('./processStreamUtilities');
 const { applicationContext } = require('../test/createTestApplicationContext');
 
@@ -310,72 +309,6 @@ describe('processStreamUtilities', () => {
       ]);
 
       expect(result.otherRecords).toEqual([mockOtherRecord]);
-    });
-  });
-
-  describe('processPractitionerMappingEntries', () => {
-    const mockPractitionerMappingEntries = [
-      mockModifyIrsPractitionerMappingRecord,
-      mockModifyPrivatePractitionerMappingRecord,
-    ];
-
-    it('should do nothing when no practitioner mapping records are found', async () => {
-      await processPractitionerMappingEntries({
-        applicationContext,
-        practitionerMappingRecords: [],
-      });
-
-      expect(
-        applicationContext.getPersistenceGateway().getCaseMetadataWithCounsel,
-      ).not.toHaveBeenCalled();
-    });
-
-    it('should retrieve and index each case for each provided practitioner mapping record', async () => {
-      const docketNumberInPractitionerMapping =
-        mockModifyIrsPractitionerMappingRecord.dynamodb.NewImage.pk.S.split(
-          '|',
-        )[1];
-
-      applicationContext
-        .getPersistenceGateway()
-        .getCaseMetadataWithCounsel.mockReturnValue(mockCaseRecord);
-
-      await processPractitionerMappingEntries({
-        applicationContext,
-        practitionerMappingRecords: mockPractitionerMappingEntries,
-      });
-
-      expect(
-        applicationContext.getPersistenceGateway().getCaseMetadataWithCounsel,
-      ).toHaveBeenCalledTimes(mockPractitionerMappingEntries.length);
-      expect(
-        applicationContext.getPersistenceGateway().getCaseMetadataWithCounsel
-          .mock.calls[0][0].docketNumber,
-      ).toEqual(docketNumberInPractitionerMapping);
-      expect(
-        applicationContext.getPersistenceGateway().bulkIndexRecords,
-      ).toHaveBeenCalled();
-    });
-
-    it('should log an error and throw an exception when bulk index returns failed records', async () => {
-      applicationContext
-        .getPersistenceGateway()
-        .getCaseMetadataWithCounsel.mockReturnValue(mockCaseRecord);
-
-      applicationContext
-        .getPersistenceGateway()
-        .bulkIndexRecords.mockReturnValueOnce({
-          failedRecords: [{ id: 'failed record' }],
-        });
-
-      await expect(
-        processPractitionerMappingEntries({
-          applicationContext,
-          practitionerMappingRecords: mockPractitionerMappingEntries,
-        }),
-      ).rejects.toThrow('failed to index practitioner mapping records');
-
-      expect(applicationContext.logger.error).toHaveBeenCalled();
     });
   });
